+++ src/rules/enum_with_type.rs
@@ -1,10 +1,16 @@
 use crate::linter::{Rule, RuleResult};
-use sv_parser::{RefNode, SyntaxTree};
+use sv_parser::{NodeEvent, RefNode, SyntaxTree};
 
 pub struct EnumWithType;
 
 impl Rule for EnumWithType {
-    fn check(&self, _syntax_tree: &SyntaxTree, node: &RefNode) -> RuleResult {
+    fn check(&mut self, _syntax_tree: &SyntaxTree, event: &NodeEvent) -> RuleResult {
+        let node = match event {
+            NodeEvent::Enter(x) => x,
+            NodeEvent::Leave(_) => {
+                return RuleResult::Skip;
+            }
+        };
         match node {
             RefNode::DataTypeEnum(x) => {
                 let (_, ref a, _, _) = x.nodes;