+++ src/rules/function_with_automatic.rs
@@ -1,10 +1,16 @@
 use crate::linter::{Rule, RuleResult};
-use sv_parser::{Lifetime, RefNode, SyntaxTree};
+use sv_parser::{Lifetime, NodeEvent, RefNode, SyntaxTree};
 
 pub struct FunctionWithAutomatic;
 
 impl Rule for FunctionWithAutomatic {
-    fn check(&self, _syntax_tree: &SyntaxTree, node: &RefNode) -> RuleResult {
+    fn check(&mut self, _syntax_tree: &SyntaxTree, event: &NodeEvent) -> RuleResult {
+        let node = match event {
+            NodeEvent::Enter(x) => x,
+            NodeEvent::Leave(_) => {
+                return RuleResult::Skip;
+            }
+        };
         match node {
             RefNode::FunctionDeclaration(x) => {
                 let (_, ref a, _) = x.nodes;