+++ src/rules/generate_keyword.rs
@@ -1,10 +1,16 @@
 use crate::linter::{Rule, RuleResult};
-use sv_parser::{RefNode, SyntaxTree};
+use sv_parser::{NodeEvent, RefNode, SyntaxTree};
 
 pub struct GenerateKeyword;
 
 impl Rule for GenerateKeyword {
-    fn check(&self, _syntax_tree: &SyntaxTree, node: &RefNode) -> RuleResult {
+    fn check(&mut self, _syntax_tree: &SyntaxTree, event: &NodeEvent) -> RuleResult {
+        let node = match event {
+            NodeEvent::Enter(x) => x,
+            NodeEvent::Leave(_) => {
+                return RuleResult::Skip;
+            }
+        };
         match node {
             RefNode::GenerateRegion(_) => RuleResult::Fail,
             _ => RuleResult::Pass,