+++ src/rules/loop_variable_declaration.rs
@@ -1,10 +1,16 @@
 use crate::linter::{Rule, RuleResult};
-use sv_parser::{ForInitialization, RefNode, SyntaxTree};
+use sv_parser::{ForInitialization, NodeEvent, RefNode, SyntaxTree};
 
 pub struct LoopVariableDeclaration;
 
 impl Rule for LoopVariableDeclaration {
-    fn check(&self, _syntax_tree: &SyntaxTree, node: &RefNode) -> RuleResult {
+    fn check(&mut self, _syntax_tree: &SyntaxTree, event: &NodeEvent) -> RuleResult {
+        let node = match event {
+            NodeEvent::Enter(x) => x,
+            NodeEvent::Leave(_) => {
+                return RuleResult::Skip;
+            }
+        };
         match node {
             RefNode::ForInitialization(ForInitialization::ListOfVariableAssignments(_)) => {
                 RuleResult::Fail